requirements.txt
chromadb
google-generativeai
sentence-transformers
pypdf
streamlit